+++ src/main/java/com/easypost/easyvcr/internalutilities/Files.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Files {
+public abstract class Files {
     /**
      * Creates a file if it doesn't exist.
      *